app/requirements.txt
kafka-python
matplotlib
keyboard